+++ src/svmClassification.h
@@ -110,7 +110,7 @@ public:
               );
     
     int getNumInputs() const override;
-    std::vector<int> getWhichInputs() const override;
+    std::vector<size_t> getWhichInputs() const override;
         
 #ifndef EMSCRIPTEN
     void getJSONDescription(Json::Value &currentModel) override;
